ocagent/ExportSpan: use size 1 instead of -1 in adding to bundler
The previous code used -1 which I had misinterpreted to mean
"unknown size" when adding elements to the bundler. However,
the correct usage is to be 1 which counts each spanData as
an entry and will trigger an upload when it hits the threshold.

Also added a regression test that is also a reproducer which makes
upto 50 simultaneous connections to the frontend server that then
generates many spans hence simulating load. However, the test is for
>=Go1.11 lest we have noisy output due to running out of file descriptors.

Fixes #13