perform indefinite background reconnection attempts
Perform an indefinite background redial on failed connections
to ensure that this exporter can reconnect to the upstream
OpenCensus interceptor/agent.

This change allows environments that don't yet have the agent
deployed to start using this exporter and gradually roll-out
the agent. It helps trivial situations like an application
starting before the agent: of which it shouldn't crash
and fail to start running -- all the exported spans before
the agent is ready for connections, or at least if the
span buffer is full, will be discarded.

Added tests to ensure that this works and lock-in that
behavior.

Fixes #18