config: do not change state based off config results
Since the OpenCensus interceptor config change hasn't
yet been implemented, we shouldn't be changing the exporters'
connection state based on its results -- it'll always return
disconnected since right now it is a noop/fail-fast on the agent.

I found this by a live integration test with a high traffic app.