Avoid deadlock due to channel used for disconnect notification (#55)
The function setStateDisconnected() writes to the channel disconnectedCh, however, the function indefiniteBackgroundConnection() that reads the channel also calls setStateDisconnected(). If other routine already wrote to the channel indefiniteBackgroundConnection() gets blocked. The fix is just to attempt to write to the channel and proceed if it is already full (capacity is set to 1).